backend/requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 数据验证 - 使用更新的版本
pydantic==2.6.0
pydantic-settings==2.2.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0

# 环境变量
python-dotenv==1.0.0

# 日志
loguru==0.7.2

# 时间处理
python-dateutil==2.8.2

# 文件处理
aiofiles==23.2.1

# 向量数据库 (后续阶段使用)
# chromadb==0.4.18
# langchain==0.0.350
# openai==1.3.7
